syntax error check

+++ include/minishell.h
@@ -6,7 +6,7 @@
 /*   By: adi-nata <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/06/26 16:44:45 by adi-nata          #+#    #+#             */
-/*   Updated: 2023/09/10 20:05:29 by adi-nata         ###   ########.fr       */
+/*   Updated: 2023/09/10 22:07:57 by adi-nata         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -96,7 +96,7 @@ int			strchr_index(char *s, char c);
 
 char		**pipe_split(const char *s, char pipe);
 void		pipe_splitter(const char *s, char pipe, char **split, size_t n);
-size_t		pipe_numstr(const char *s, char pipe);
+int			pipe_numstr(const char *s, char pipe);
 
 void		signal_handler(int sig);
 void		signal_print(int sig);
